Rename controllers
